test: authReducer
